hack(test): tsx 119
Hack to make the CI pass. I'm not sure why the error is happening, but I
think it may have something to do with the nightly version of Neovim.
Before updating the docker image on my local machine, all test where
passing and after updating the test for 119 for tsx started failling.