feat(command): support multiple inputs
The `:Refactor` command can now receive multiple inputs (i. e.
extract_to_file needs a file name and the name of the function being
extracted as input).

Currently, there is a problem with how the windows are managed when
extracting code to a new file (the preview produces side effects even if
the command is canceled). So, the preview is currently disabled for
refactors that extract code to a new file.